@@ -1,7 +1,7 @@
 #ifndef __KSU_H_APK_V2_SIGN
 #define __KSU_H_APK_V2_SIGN
 
-#include "linux/types.h"
+#include <linux/types.h>
 
 bool is_manager_apk(char *path);
 
