@@ -1,8 +1,8 @@
 #ifndef __KSU_H_KSU_MANAGER
 #define __KSU_H_KSU_MANAGER
 
-#include "linux/cred.h"
-#include "linux/types.h"
+#include <linux/cred.h>
+#include <linux/types.h>
 
 #define KSU_INVALID_UID -1
 
@@ -32,7 +32,5 @@ static inline void ksu_invalidate_manager_uid()
 {
 	ksu_manager_uid = KSU_INVALID_UID;
 }
-
-bool become_manager(char *pkg);
 
 #endif
