@@ -1,7 +1,7 @@
 #ifndef __KSU_H_SEPOLICY
 #define __KSU_H_SEPOLICY
 
-#include "linux/types.h"
+#include <linux/types.h>
 
 #include "ss/policydb.h"
 
