@@ -2,10 +2,10 @@ package me.weishu.kernelsu.ui.component.profile
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ReadMore
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material.icons.filled.ArrowDropUp
 import androidx.compose.material.icons.filled.Create
-import androidx.compose.material.icons.filled.ReadMore
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExposedDropdownMenuBox
@@ -105,7 +105,7 @@ fun TemplateConfig(
                             IconButton(onClick = {
                                 onViewTemplate(tid)
                             }) {
-                                Icon(Icons.Filled.ReadMore, null)
+                                Icon(Icons.AutoMirrored.Filled.ReadMore, null)
                             }
                         }
                     )
