@@ -83,8 +83,11 @@ class SuperUserViewModel : ViewModel() {
 
     val appList by derivedStateOf {
         sortedList.filter {
-            it.label.contains(search) || it.packageName.contains(search) || HanziToPinyin.getInstance()
-                .toPinyinString(it.label).contains(search)
+            it.label.contains(search, true) || it.packageName.contains(
+                search,
+                true
+            ) || HanziToPinyin.getInstance()
+                .toPinyinString(it.label).contains(search, true)
         }.filter {
             it.uid == 2000 // Always show shell
                     || showSystemApps || it.packageInfo.applicationInfo.flags.and(ApplicationInfo.FLAG_SYSTEM) == 0
@@ -104,7 +107,7 @@ class SuperUserViewModel : ViewModel() {
             }
         }
 
-        val intent = Intent(ksuApp, KsuService::class.java);
+        val intent = Intent(ksuApp, KsuService::class.java)
 
         val task = KsuService.bindOrTask(
             intent,
@@ -116,7 +119,7 @@ class SuperUserViewModel : ViewModel() {
     }
 
     private fun stopKsuService() {
-        val intent = Intent(ksuApp, KsuService::class.java);
+        val intent = Intent(ksuApp, KsuService::class.java)
         KsuService.stop(intent)
     }
 
