@@ -82,7 +82,7 @@ async def main():
         exit(1)
     print("[+] Logging in Telegram with bot")
     script_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
-    session_dir = os.path.join(script_dir, "ksubot.session")
+    session_dir = os.path.join(script_dir, "ksubot")
     async with await TelegramClient(session=session_dir, api_id=API_ID, api_hash=API_HASH).start(bot_token=BOT_TOKEN) as bot:
         caption = [""] * len(files)
         caption[-1] = get_caption()
