@@ -22,7 +22,7 @@ export default defineConfig({
     ],
 
     footer: {
-        message: 'Lançado sob a Licença GPL3.',
+        message: 'Lançado sob a Licença GPL3',
         copyright: 'Copyright © Desenvolvedores do KernelSU atuais de 2022'
     },
 
@@ -47,7 +47,7 @@ function sidebarGuide() {
           { text: 'O que é KernelSU?', link: '/pt_BR/guide/what-is-kernelsu' },
           { text: 'Diferença com Magisk', link: '/pt_BR/guide/difference-with-magisk' },
           { text: 'Instalação', link: '/pt_BR/guide/installation' },
-          { text: 'Como construir?', link: '/pt_BR/guide/how-to-build' },
+          { text: 'Como compilar?', link: '/pt_BR/guide/how-to-build' },
           { text: 'Integração para dispositivos não GKI', link: '/pt_BR/guide/how-to-integrate-for-non-gki'},
           { text: 'Dispositivos com suporte não oficial', link: '/pt_BR/guide/unofficially-support-devices.md' },
           { text: 'Guias de módulo', link: '/pt_BR/guide/module.md' },
